# #12116 Speed up http_headers operations