--- a/src/twisted/web/newsfragments/12116.feature
+++ b/src/twisted/web/newsfragments/12116.feature
@@ -1,0 +1,1 @@
+twisted.web.http_headers now uses less CPU.