## Scope and purpose

Fixes #12116

The most significant changes are:

1. Change the internal representation of headers, so they're canonicalized once instead of lowercased incoming and canonicalized outgoing.
2. Cache canonicalization, with limits to prevent high memory usage by attackers.
3. Get rid of runtime type checks, given static type checking is now available in Python.

Same benchmark as #12108, measuring sending and parsing a HTTP client request. This will likely speed up the HTTP server as well.

Before:

```
Benchmark 1: python http_client_benchmark.py
  Time (mean ± σ):      3.262 s ±  0.027 s    [User: 3.250 s, System: 0.012 s]
  Range (min … max):    3.223 s …  3.307 s    10 runs
```

After:

```
Benchmark 1: python http_client_benchmark.py
  Time (mean ± σ):      2.929 s ±  0.036 s    [User: 2.918 s, System: 0.011 s]
  Range (min … max):    2.877 s …  3.016 s    10 runs
```