@adiroiban 2024-03-18
Just a FYI. We have this PR https://github.com/twisted/twisted/pull/12094 to look into increasing the limit for headers

I didn't had time to follow up with it.

@itamarst  or @glyph  maybe you can take a look and see if we can merge it

I am +1 for Itamar
I am ok to have hardcoded limits, as long as they can handle all real world usage.